add utils and refeact code
